Increase pixel light count
To keep render cost low, Unity sets the base pixel light count to
only 4. Bump this up to 8 because some some areas weren't being
lit correctly, but this should be a temporary fix. Leaving this
high will be a large performance hit (100 lights on an object
is as heavy as 100 of that one object), so doing something like
combining a group of mushroom's light source into a single point
would be a better option.